+++ golem-rib-repl/Cargo.toml
@@ -14,19 +14,26 @@ path = "src/lib.rs"
 name = "golem-rib-repl"
 path = "src/main.rs"
 
+[features]
+default = []
+bin = ["golem-test-framework", "golem-worker-executor-base", "golem-service-base", "golem-api-grpc"]
 
 [dependencies]
 golem-wasm-rpc = { path = "../wasm-rpc", version = "=0.0.0", default-features = false, features = ["host"] }
-golem-common = { path = "../golem-common", version = "=0.0.0" }
-golem-worker-executor-base = { path = "../golem-worker-executor-base", version = "=0.0.0"}
-golem-test-framework = { path = "../golem-test-framework", version = "=0.0.0"}
-golem-service-base = { path = "../golem-service-base", version = "=0.0.0"}
+
 golem-rib = { path = "../golem-rib", version = "=0.0.0" }
 golem-wasm-ast = { path = "../wasm-ast", version = "=0.0.0", default-features = false, features = [
     "analysis",
     "bincode",
 ] }
-golem-api-grpc = { path = "../golem-api-grpc", version = "=0.0.0"}
+golem-common = { path = "../golem-common", version = "=0.0.0" }
+
+
+# Binary only dependencies
+golem-api-grpc = { path = "../golem-api-grpc", version = "=0.0.0", optional = true}
+golem-service-base = { path = "../golem-service-base", version = "=0.0.0", optional = true}
+golem-test-framework = { path = "../golem-test-framework", version = "=0.0.0", optional = true}
+golem-worker-executor-base = { path = "../golem-worker-executor-base", version = "=0.0.0", optional = true}
 
 anyhow = { workspace = true }
 async-trait = { workspace = true }
